rag-service/requirements.txt
python-dotenv
faiss-cpu
google-generativeai
fastapi[standard]
pydantic
langchain-community
langchain-huggingface
sentence_transformers
pandas